+++ cmds/command_shows.go
@@ -25,13 +25,14 @@ var (
 	_showsResetAt       = ShowsCmd.Flag.String("reset_at", "", consts.ResetAtUsage)
 	_showsReason        = ShowsCmd.Flag.String("r", cfg.DefaultConfig().Reason, consts.ReasonUsage)
 	_showsMessage       = ShowsCmd.Flag.String("message", cfg.DefaultConfig().Msg, consts.ReportMsgUsage)
+	_showsLinks         = ShowsCmd.Flag.String("links", consts.EmptyString, consts.ShowsLinksUsage)
 
 	validShowsActions = []string{
 		"trending", "popular", "favorited", "played", "watched", "collected",
 		"anticipated", "boxoffice", "updates", "updated_ids", "summary", "aliases", "certifications",
 		"collection_progress", "watched_progress", "releases", "translations", "comments", "lists", "people", "ratings",
 		"releated", "stats", "studios", "watching", "next_episode", "last_episode", "videos", "refresh",
-		consts.Report, consts.Sentiments}
+		consts.Report, consts.Sentiments, consts.WatchNow, consts.JustwatchLinks}
 )
 
 // ShowsCmd returns movies and episodes that a user has watched, sorted by most recent.
@@ -91,6 +92,9 @@ func showsFunc(cmd *Command, _ ...string) error {
 
 		consts.Report:     handlers.ShowsReportHandler{},
 		consts.Sentiments: handlers.ShowsSentimentsHandler{},
+
+		consts.WatchNow:       handlers.ShowsWatchNowHandler{},
+		consts.JustwatchLinks: handlers.ShowsJustwatchLinksHandler{},
 	}
 	handler, err = cmd.common.GetHandlerForMap(options.Action, allHandlers)